+++ packages/bpk-component-label/src/BpkLabel-test.js
@@ -4,9 +4,12 @@ import BpkLabel from './BpkLabel'
 
 describe('BpkLabel', () => {
   it('should render correctly', () => {
-    const tree = renderer.create(
-      <BpkLabel label='Origin' htmlFor='origin' />
-    ).toJSON()
+    const tree = renderer.create(<BpkLabel label='Origin' htmlFor='origin' />).toJSON()
+    expect(tree).toMatchSnapshot()
+  })
+
+  it('should render correctly with a "white" attribute', () => {
+    const tree = renderer.create(<BpkLabel label='Origin' htmlFor='origin' white />).toJSON()
     expect(tree).toMatchSnapshot()
   })
 })